+++ tests/lib/albatross/test/test_models.h
@@ -19,7 +19,7 @@
 #include <albatross/SparseGP>
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace albatross {
