+++ tests/lib/albatross/test/test_serialize.cc
@@ -18,9 +18,9 @@
 
 #include <gtest/gtest.h>
 
-#include "test_models.h"
+#include "albatross/test/test_models.h"
 #include "test_serialize.h"
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace cereal {
