+++ tests/test_callers.cc
@@ -14,7 +14,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test_covariance_utils.h"
+#include "albatross/test/test_covariance_utils.h"
 
 namespace albatross {
