+++ tests/test_conditional_gaussian.cc
@@ -12,7 +12,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test_models.h"
+#include "albatross/test/test_models.h"
 
 namespace albatross {
