+++ tests/test_covariance_function.cc
@@ -12,7 +12,7 @@
 
 #include <albatross/CovarianceFunctions>
 
-#include "test_covariance_utils.h"
+#include "albatross/test/test_covariance_utils.h"
 #include <gtest/gtest.h>
 
 namespace albatross {
