+++ tests/test_cross_validation.cc
@@ -14,7 +14,7 @@
 #include <chrono>
 #include <gtest/gtest.h>
 
-#include "test_models.h"
+#include "albatross/test/test_models.h"
 
 namespace albatross {
