+++ tests/test_evaluate.cc
@@ -12,7 +12,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 #include <albatross/Evaluation>
 
 namespace albatross {
