+++ tests/test_model_adapter.cc
@@ -11,7 +11,7 @@
  */
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace albatross {
