+++ tests/test_parameter_handling_mixin.cc
@@ -12,7 +12,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace albatross {
