+++ tests/test_radial.cc
@@ -14,7 +14,7 @@
 #include <array>
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace albatross {
