+++ tests/test_samplers.cc
@@ -18,7 +18,7 @@
 
 #include <fstream>
 
-#include "test_models.h"
+#include "albatross/test/test_models.h"
 
 namespace albatross {
