+++ tests/test_scaling_function.cc
@@ -13,7 +13,7 @@
 #include <albatross/Evaluation>
 #include <gtest/gtest.h>
 
-#include "test_utils.h"
+#include "albatross/test/test_utils.h"
 
 namespace albatross {
