+++ tests/test_sparse_gp.cc
@@ -12,7 +12,7 @@
 
 #include <gtest/gtest.h>
 
-#include "test_models.h"
+#include "albatross/test/test_models.h"
 #include <chrono>
 
 #include <albatross/SparseGP>
